requirements.txt
astroid==1.5.3
autopep8==1.3.2
backports.functools-lru-cache==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
configparser==3.5.0
enum34==1.1.6
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
peewee==2.10.1
pycodestyle==2.3.1
pylint==1.7.4
PyMySQL==0.7.11
python-dotenv==0.7.1
requests==2.18.4
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.22
wrapt==1.10.11
